requirements.txt
aiogram==3.4.1
aiosqlite==0.19.0
python-dotenv==1.0.0
aiofiles==23.2.1
pillow==10.2.0
requests==2.31.0